liver_system/requirements.txt
Flask_Migrate==2.4.0
scikit_image==0.14.2
WTForms==2.2.1
alembic==1.0.9
Flask_Login==0.5.0
SQLAlchemy==1.3.1
Flask==1.0.2
Werkzeug==0.14.1
Flask_Bootstrap==3.3.7.1
numpy==1.16.2
nibabel==3.1.0
Flask_Dropzone==1.5.4
Flask_Bcrypt==0.7.1
Flask_SQLAlchemy==2.3.2
matplotlib==3.0.3
Flask_Script==2.0.6
Flask_RESTful==0.3.7
Flask_WTF==0.14.2
imageio==2.8.0
Pillow==7.1.2
pydicom==2.0.0
torch==1.5.1
torchvision==0.6.1